test(evaluator): split engine/tests/mod.rs into themed child modules
Issue #391 item 1: extract the oversized src/evaluator/engine/tests/mod.rs
(3799 -> 263 lines) into 14 new themed child test modules, continuing the
existing engine/tests themed-file pattern. Pure test code motion: all 159
evaluator::engine::tests functions unchanged (same count before and after);
mod.rs retains only shared helpers and mod declarations.

No behavior change: full suite green (1608), clippy/fmt clean both feature sets.
